pyproject.toml
[project]
name = "repaso-fundepos-modelacion-fin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas-profiling>=3.6.6",
    "prefect>=3.6.25",
    "setuptools>=82.0.1",
    "ydata-profiling>=4.18.1",
    "yfinance>=1.2.1",
]
